feat: Add stateTool error handling && add client powerDisplay
# stateTool error handling
Power wasn't showing from the stateTool when the president was shown.
This ended up being caused by the powerActive not being set.
Added checks on the stateTool fetch request to pick up server errors

# powerDisplay
Make a function in the game client that will eventually show the appropriate
buttons/cards for the president to complete the powers